Fixing #282 - using Semantic Kernel memory for vectors
This moves from the custom solution for vector searching on Postgres, instead using the SK memory feature to do it. Added some new SK dependencies for the memory (and PG memory store), then removed the Embedding column from the current data model, as there is a new table with all that in it.

Refactored the seed logic to load the memory store using the previously generated embeddings.

Changed the CatalogAPI route to use the ISemanticTextMemory search feature to search memory, rather than the custom SQL query. This does mean we don't get distance surfaced, also, pagination is currently lost and SK memory doesn't support that (we could roll that ourselves if we want).

Included a fix so that AOAI can be deployed (issue #280), and pgadmin for easier debugging of the data in the database.